changes to avoid public api changes
